Reorganise tests.
